added ratings and reviews routes

+++ src/controllers/v1/reviewControllers.ts
@@ -1,15 +1,12 @@
-import { Request, Response } from "express";
+import { Response } from "express";
 import { prisma } from "../../db";
+import { CustomRequest } from "../../types";
 import { ApiResponse } from "../../utils/APIResponse";
 import { asyncHandler } from "../../utils/asyncHandler";
 
-interface CustomRequest extends Request {
-  user?: any; // You can replace 'any' with a proper User type
-}
-
 export const createReview = asyncHandler(
   async (req: CustomRequest, res: Response) => {
-    const { bookId, rating, review } = req.body;
+    const { bookId, review } = req.body;
     const userId = req.user.id;
 
     const book = await prisma.book.findUnique({
@@ -36,17 +33,52 @@ export const createReview = asyncHandler(
   }
 );
 
-
 export const getReviews = asyncHandler(
   async (req: CustomRequest, res: Response) => {
-    const bookId = req.params.bookId;
+    const bookId = req.query.bookId as string;
 
     const reviews = await prisma.review.findMany({
       where: { bookId },
-      include: { user: true },
+      take: req.query.limit ? Number(req.query.limit) : 10,
+      skip: req.query.skip ? Number(req.query.skip) : 0,
     });
 
     res.status(200).json(new ApiResponse(200, reviews, "Reviews retrieved"));
+  }
+);
+
+export const getReviewById = asyncHandler(
+  async (req: CustomRequest, res: Response) => {
+    const reviewId = req.params.reviewId;
+    console.log(reviewId);
+
+    const userId = req.user.id;
+    const review = await prisma.review.findUnique({
+      where: { id: reviewId, userId },
+    });
+    if (!review) {
+      return res
+        .status(404)
+        .json(new ApiResponse(404, null, "Review not found"));
+    }
+    res.status(200).json(new ApiResponse(200, review, "Review retrieved"));
+
+    // You can also use the following query for getting a review by its ID and user ID
+    // const review = await prisma.review.findUnique({
+    //   where: { id: reviewId },
+    // });
+    // if (!review) {
+    //   return res.status(404).json({
+    //     success: false,
+    //     error: "Review not found",
+    //   });
+    // }
+    // if (review.userId!== userId) {
+    //   return res.status(403).json({
+    //     success: false,
+    //     error: "Unauthorized to access this review",
+    //   });
+    // }
   }
 );
 
@@ -54,9 +86,10 @@ export const updateReview = asyncHandler(
   async (req: CustomRequest, res: Response) => {
     const reviewId = req.params.reviewId;
     const { review } = req.body;
+    const userId = req.user.id;
 
     const updatedReview = await prisma.review.update({
-      where: { id: reviewId },
+      where: { id: reviewId, userId },
       data: { review },
     });
 
@@ -64,19 +97,27 @@ export const updateReview = asyncHandler(
       .status(200)
       .json(new ApiResponse(200, updatedReview, "Review updated successfully"));
   }
-
-
-
 );
 
 export const deleteReview = asyncHandler(
   async (req: CustomRequest, res: Response) => {
     const reviewId = req.params.reviewId;
+    const userId = req.user.id;
+
+    const review = await prisma.review.findUnique({
+      where: { id: reviewId, userId },
+    });
+
+    if (!review) {
+      return res
+        .status(404)
+        .json(new ApiResponse(404, null, "Review not found"));
+    }
 
     await prisma.review.delete({
       where: { id: reviewId },
     });
 
-    res.status(204).json(new ApiResponse(204, undefined, "Review deleted"));
+    res.status(200).json(new ApiResponse(200, undefined, "Review deleted"));
   }
 );
